Оптимизация рефакторинг
