+++ src/main/java/lab_3/ReadData.java
@@ -1,4 +1,4 @@
-package Lab3;
+package lab_3;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
